Custom metric code and promt based
