+++ sdk/ai/azure-ai-generative/azure/ai/generative/evaluate/_metric_handler.py
@@ -71,7 +71,7 @@ class MetricHandler(object):
         return metrics_data
 
     def calculate_metrics(self):
-        LOGGER.info(f"Calculating inbuilt metrics : {[metric for metric in self.metrics]}")
+        LOGGER.info(f"Calculating builtin metrics : {[metric for metric in self.metrics]}")
         from azureml.metrics import compute_metrics
 
         metrics_calculation_data = self._get_data_for_metrics()
@@ -94,91 +94,7 @@ class MetricHandler(object):
         return metrics_value
 
 
-class CodeMetricHandler(MetricHandler):
-    def __init__(
-            self,
-            task_type,
-            prediction_data,
-            test_data,
-            truth_data=None,
-            prediction_data_column_name=None,
-            ground_truth_column_name=None,
-            metrics_mapping=None,
-            metrics=None,
-            type_to_kwargs=None,
-    ):
 
-        super().__init__(
-            task_type=task_type,
-            prediction_data=prediction_data,
-            test_data=test_data,
-            truth_data=truth_data,
-            prediction_data_column_name=prediction_data_column_name,
-            ground_truth_column_name=ground_truth_column_name,
-            metrics_mapping=metrics_mapping,
-            metrics=metrics,
-            type_to_kwargs=type_to_kwargs,
-        )
-
-        self._validate()
-
-    def _validate(self):
-        supported_list = [isinstance(metric, CodeMetric) for metric in self.metrics]
-        if not all(supported_list):
-            raise Exception(f"{self.__class__.__name__} supports only {CodeMetric.__class__.__name__} type of metrics")
-
-    def calculate_metrics(self):
-        LOGGER.info(f"Calculating code metrics : {[metric.name for metric in self.metrics]}")
-        metrics_dict = {"artifacts": {}, "metrics": {}}
-        metric_results_futures = {}
-        with tqdm(total=len(self.metrics)) as progress_bar:
-            with ThreadPoolExecutor(thread_name_prefix="code_metrics") as thread_pool:
-                for metric in self.metrics:
-                    metric_values = []
-                    metric_results_futures.update({metric.name: thread_pool.submit(
-                        self._calculate_metric, metric, self.test_data.to_dict('records'),
-                        self.prediction_data.to_dict('records') if self.prediction_data is not None else None
-                    )})
-
-                for metric_name, metric_result_future in metric_results_futures.items():
-                    try:
-                        metric_result = metric_result_future.result()
-                        metrics_dict["artifacts"].update(metric_result["artifacts"])
-                        if "metrics" in metric_result.keys() and metric_result["metrics"] is not None:
-                            metrics_dict["metrics"].update(metric_result["metrics"])
-                        progress_bar.update(1)
-                    except Exception as ex:
-                        progress_bar.update(1)
-                        LOGGER.info(f"Error calculating value for {metric_name}, failed with error {str(ex)} : Stack trace : {str(ex.__traceback__)}")
-
-        return metrics_dict
-
-    def _calculate_metric(self, metric, data, response):
-
-        row_metric_futures = []
-        row_metric_result = []
-        aggregated_metrics = None
-
-        with ThreadPoolExecutor(thread_name_prefix="code_metrics_row") as thread_pool:
-            for i in range(0, len(data)):
-                row_metric_futures.append(thread_pool.submit(
-                    metric.calculate, data[i], response[i]
-                ))
-
-            for row_metric_future in row_metric_futures:
-                try:
-                    row_metric_result.append(row_metric_future.result())
-                except Exception as ex:
-                    LOGGER.info(f"Error calculating value for a row for metric {metric.name} , failed with error {str(ex)} : Stack trace : {str(ex.__traceback__)}")
-                    row_metric_result.append(NaN)
-
-        if metric.aggregator:
-            aggregated_metrics = metric.aggregator(row_metric_result)
-
-        return {
-            "artifacts": {metric.name: row_metric_result},
-            "metrics": aggregated_metrics
-        }
